Backend/OCR/requirements.txt
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
nms==0.1.6
numpy==1.15.0
opencv-python==4.1.1.26
# opencv-text-detection==0.0.1
Pillow==6.2.0
pybase64==1.0.0
pytesseract==0.3.0
Werkzeug==0.16.0
tesseract==0.1.3